Fix crash when interface without IFLA_INFO_KIND is present
When an interface is present that has a IFLA_LINKINFO nested attribute
block, but without that block then having IFLA_INFO_KIND, then a
null-dereference occurs. An example of such an interface is a pimreg
interface created for a VRF.

Fix this by checking for the presence of the IFLA_INFO_KIND attribute
before dereferencing it, and bail early since this is expected and it
isn't "one of ours".

Fixes: VRVDR-53084